Final Update?
Refactored some functions and fields to improve legibility. Replaced functions using spigot's async thread with its own to improve performance. Removed unnecessary thread synchronization with update utility functions. Added a fallback (legacy) version of the block update functions incase the custom one does not work for whatever reason in the future or with versions not tested. Added some small general improvements to the build configuration and plugin commands. Ensured this version was built with JDK 1.8 instead of 17 which has caused problems for players running on earlier versions of the game.